magic tree-watcher to auto-call accessor init
Adds support for magical "__foobar_accessor_ready" methods. Any
parameterless method which has name starting with two underscores ('__')
and a postfix "_accessor_ready" is considered an accessor initializer.

Accessor initializer can be used to do anything the script normally would
do in _ready(). This solves the issue of multiple traits not being able
to define a _ready() -function, due to shadowing.

Automatically registers an autoload when enabling the component plugin.
This autoload globally watches the tree for any added nodes. The
autoload then attempts to find any methods that look like component
accessor _ready -methods, and calls them.

This probably has a significant performance impact, given this has to be
performed on every single node that is added to the tree. To alleviate
this a bit, the autoload maintains a cached dictionary, holding a lookup
table for quickly determining a list of init method names for a script
UID. Additionally, the autoload attempts to return early, whenever
possible.